Revert "Merge pull request #3 from ArduPilot/master"
This reverts commit 13c39969d86c5ae35965284f3427071daf5678c7.

+++ libraries/AP_RSSI/AP_RSSI.cpp
@@ -219,7 +219,7 @@ void AP_RSSI::check_pwm_pin_rssi()
         // failed to attach interrupt
         gcs().send_text(MAV_SEVERITY_WARNING,
                         "RSSI: Failed to attach to pin %u",
-                        (unsigned int)rssi_analog_pin);
+                        rssi_analog_pin);
         return;
     }
 }
